day5 part1 solution

+++ src/IDRange.java
@@ -1,14 +1,14 @@
 
 public class IDRange {
-    int low = 0;
-    int high = 0;
+    long low = 0;
+    long high = 0;
 
-    public IDRange(int low, int high) {
+    public IDRange(long low, long high) {
         this.low = low;
         this.high = high;
     }
 
-    public boolean contains(int id) {
+    public boolean contains(long id) {
         return (id >= low && id <= high);
     }
